+++ src/components/BulkEdit.test.js
@@ -137,6 +137,36 @@ describe('BulkEdit', () => {
     expect(identifierButton).toHaveAttribute('class', 'button primary');
   });
 
+  it('should change display text area on query tab ', async () => {
+    renderBulkEdit();
+
+    const queryButton = screen.getByRole('button', { name: /list.filters.query/ });
+
+    userEvent.click(queryButton);
+
+    const textArea = screen.getByRole('textbox');
+    const searchButton = screen.getByRole('button', { name: /textArea.search/ });
+    const resetButton = screen.getByTestId('reset-button');
+
+    expect(queryButton).toHaveAttribute('class', 'button primary');
+
+    expect(textArea).toBeVisible();
+
+    expect(searchButton).toBeDisabled();
+    expect(resetButton).toBeDisabled();
+
+    await userEvent.type(textArea, '12', { delay: 10 });
+
+    expect(textArea).toHaveValue('12');
+
+    expect(searchButton).toBeEnabled();
+    expect(resetButton).toBeEnabled();
+
+    userEvent.click(resetButton);
+
+    expect(textArea).not.toHaveValue();
+  });
+
   it('should display select', () => {
     renderBulkEdit();