+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 * [UIBULKED-11](https://issues.folio.org/browse/UIBULKED-11) Bulk Edit app - permissions 
 * [UIBULKED-19](https://issues.folio.org/browse/UIBULKED-19) Download errors (CSV)
 * [UIBULKED-17](https://issues.folio.org/browse/UIBULKED-17) Start bulk edit
+* [UIBULKED-15](https://issues.folio.org/browse/UIBULKED-15) Add Query textarea